pyproject.toml
[project]
name = "wuerth_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "grandalf>=0.8",
    "jupyterlab>=4.4.3",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.5",
    "langchain-openai>=0.3.21",
    "langgraph==0.3.1",
    "langgraph-cli[inmem]>=0.1.0,<0.3.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5,<3.0.0",
    "pyparsing==3.1.1",
    "python-dotenv>=1.1.0",
    "rdflib==7.0.0",
]
